LayoutTests/imported/w3c/web-platform-tests/css/css-mixins/contents-nested-declarations-fallback-expected.txt


FAIL Can mix declarations into pseudo-elements via @contents assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
PASS Nested declarations from @contents have top-level specificity behavior
